requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.6
colorama==0.4.4
confusable-homoglyphs==3.2.0
coverage==5.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-autoslug==1.9.8
django-heroku==0.3.1
django-registration==3.1.1
django-stdimage==5.3.0
Faker==7.0.1
filelock==3.0.12
gunicorn==20.1.0
iniconfig==1.1.1
lxml==4.6.3
mixer==7.1.2
model-bakery==1.2.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
progressbar2==3.53.1
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
python-dateutil==2.8.1
python-dotenv==0.17.0
python-utils==2.5.6
pytils==0.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
whitenoise==5.2.0
